+++ package.json
@@ -37,6 +37,7 @@
     "cmdk": "^1.0.0",
     "date-fns": "^3.6.0",
     "embla-carousel-react": "^8.1.6",
+    "googleapis": "^140.0.1",
     "lucide-react": "^0.400.0",
     "next": "14.2.4",
     "react": "^18",
